Next level
- Rename all appropriate occurrences of "world" to "level"
- Simplify uniform buffers
- Fix indirect cull dispatch only binding frame uniforms
- Fix level renderer reload resetting debug mode
- Properly register backend argument

+++ src/main/java/com/jozufozu/flywheel/backend/engine/uniform/UniformBuffer.java
@@ -1,5 +1,6 @@
 package com.jozufozu.flywheel.backend.engine.uniform;
 
+import org.jetbrains.annotations.Nullable;
 import org.lwjgl.opengl.GL32;
 
 import com.jozufozu.flywheel.backend.gl.buffer.GlBuffer;
@@ -7,35 +8,53 @@ import com.jozufozu.flywheel.backend.gl.buffer.GlBufferUsage;
 import com.jozufozu.flywheel.lib.math.MoreMath;
 import com.jozufozu.flywheel.lib.memory.MemoryBlock;
 
-public class UniformBuffer<T extends UniformProvider> {
+public class UniformBuffer {
 	private final int index;
-	private final GlBuffer buffer;
-	public final T provider;
-	private final MemoryBlock data;
+	private final MemoryBlock clientBuffer;
+	@Nullable
+	private GlBuffer buffer;
+	private boolean needsUpload;
 
-	public UniformBuffer(int index, T provider) {
+	public UniformBuffer(int index, int size) {
 		this.index = index;
-		this.buffer = new GlBuffer(GlBufferUsage.DYNAMIC_DRAW);
-		this.provider = provider;
 
 		// renderdoc complains if the size of the buffer is not 16-byte aligned,
 		// though things work fine on my machine if we're short a few bytes
-		var actualBytes = MoreMath.align16(provider.byteSize());
-		this.data = MemoryBlock.mallocTracked(actualBytes);
-		this.data.clear();
+		clientBuffer = MemoryBlock.malloc(MoreMath.align16(size));
+		clientBuffer.clear();
 	}
 
-	public void update() {
-		provider.write(data.ptr());
-		buffer.upload(data);
+	public long ptr() {
+		return clientBuffer.ptr();
+	}
+
+	public void markDirty() {
+		needsUpload = true;
+	}
+
+	public void clear() {
+		clientBuffer.clear();
+		markDirty();
 	}
 
 	public void bind() {
-		GL32.glBindBufferRange(GL32.GL_UNIFORM_BUFFER, index, buffer.handle(), 0, data.size());
+		if (buffer == null) {
+			buffer = new GlBuffer(GlBufferUsage.DYNAMIC_DRAW);
+			needsUpload = true;
+		}
+
+		if (needsUpload) {
+			buffer.upload(clientBuffer);
+			needsUpload = false;
+		}
+
+		GL32.glBindBufferRange(GL32.GL_UNIFORM_BUFFER, index, buffer.handle(), 0, clientBuffer.size());
 	}
 
 	public void delete() {
-		data.free();
-		buffer.delete();
+		if (buffer != null) {
+			buffer.delete();
+			buffer = null;
+		}
 	}
 }
